requirements.txt
GitPython >= 3.1,
sh >= 1.14

